refactor: improve naming
