+++ src/code-components/Router/Router.tsx
@@ -2,10 +2,9 @@ import {
   DataProvider,
   GlobalActionsProvider,
 } from "@plasmicapp/react-web/lib/host";
-import { ReactNode, useMemo } from "react";
-import { Query } from "./utils/queryString";
+import { ReactNode } from "react";
 import { RouteStorageType, useStorage } from "./storage/useStorage";
-import { useQuery } from "./useQuery";
+import { RouterActions, useRouterContext } from "./useRouterContext";
 
 export interface RouterProps {
   initialQueryString?: string;
@@ -13,51 +12,20 @@ export interface RouterProps {
   children: ReactNode;
 }
 
-export interface RouteContext extends RouterActions {
-  query: Query;
-}
-
-export interface RouterActions {
-  setQuery(
-    query: Record<string, string | null>,
-    options?: {
-      /**
-       * @default true
-       */
-      merge?: boolean;
-      /**
-       * @default false
-       */
-      push?: boolean;
-    },
-  ): void;
-}
-
-export function Router({ initialQueryString, storage, children }: RouterProps) {
-  const storageRef = useStorage({ initialQueryString, type: storage });
-  const [query, setQuery] = useQuery(storageRef);
-
-  const { context, actions } = useMemo<{
-    context: RouteContext;
-    actions: RouterActions;
-  }>(() => {
-    return {
-      context: {
-        query,
-        setQuery,
-      },
-      actions: {
-        setQuery,
-      },
-    };
-  }, [query, setQuery]);
+export function Router({
+  initialQueryString,
+  storage: storageType,
+  children,
+}: RouterProps) {
+  const storage = useStorage({ initialQueryString, type: storageType });
+  const { route, actions } = useRouterContext(storage);
 
   return (
     <GlobalActionsProvider
       contextName="Router"
       actions={actions as Record<keyof RouterActions, Function>}
     >
-      <DataProvider name="route" data={context}>
+      <DataProvider name="route" data={route}>
         {children}
       </DataProvider>
     </GlobalActionsProvider>
